Change ENR sequence number to 64 bits
Here is the ENR spec that mentions 64-bit:
https://github.com/ethereum/devp2p/blob/6eddaf50298d551a83bcc242e7ce7024c6cc8590/enr.md

+++ ddht/v5_1/alexandria/sedes.py
@@ -1,4 +1,4 @@
-from ssz.sedes import Container, List, uint8, uint16, uint32, uint256
+from ssz.sedes import Container, List, uint8, uint16, uint64, uint256
 
 
 class ByteList(List):  # type: ignore
@@ -15,8 +15,8 @@ class ByteList(List):  # type: ignore
 byte_list = ByteList(max_length=2048)
 
 
-PingSedes = Container(field_sedes=(uint32, uint256))
-PongSedes = Container(field_sedes=(uint32, uint256))
+PingSedes = Container(field_sedes=(uint64, uint256))
+PongSedes = Container(field_sedes=(uint64, uint256))
 
 FindNodesSedes = Container(field_sedes=(List(uint16, max_length=256),))
 FoundNodesSedes = Container(field_sedes=(uint8, List(byte_list, max_length=32)))
